requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.43.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
fake-useragent==1.4.0
tenacity==8.2.3
loguru==0.7.2
rich==13.7.0
python-dateutil==2.8.2
apscheduler==3.10.4
flask==3.0.2

# v2.0
anthropic>=0.25.0
pyyaml==6.0.1
jinja2==3.1.4

# v3.0 — FastAPI + multi-agents
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
